requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
openpyxl==3.1.2
pandas==2.1.4
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
